Refactored InsertPreparedStatement to a separate class.
